+++ src/constants.js
@@ -11,6 +11,7 @@ const UE4SS_MODS_PATH = path.join(WIN64_PATH, 'ue4ss', 'Mods');
 const GAME_EXECUTABLE = path.join(WIN64_PATH, 'ForeverWinter-Win64-Shipping.exe');
 
 const SIGNATURE_BYPASS_URL = 'https://www.nexusmods.com/theforeverwinter/mods/57';
+const UE4SS_RELEASES_URL = 'https://github.com/UE4SS-RE/RE-UE4SS/releases';
 
 const MOD_TYPES = {
   PAKS_MODS: 'tfw-paks-mods',
@@ -29,6 +30,7 @@ module.exports = {
   PAKS_ROOT_PATH,
   SIGNATURE_BYPASS_URL,
   STEAM_APP_ID,
+  UE4SS_RELEASES_URL,
   UE4SS_MODS_PATH,
   WIN64_PATH,
 };